In the middle of a refactor. Still needs testing. Summary of changes below
removed many number formatting messages that werent used
Removed some fields that were being calculated that weren't actually being used to calculate anything. In the process of standardizing variable names and inputs across the histogram plots for the eventual goal of having a standardized plotting class that can be subclassed for particular plots.